try removing more and running migrate:reset

+++ db/schema.rb
@@ -15,8 +15,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_11_26_182903) do
   enable_extension "pg_catalog.plpgsql"
 
   create_table "screeners", force: :cascade do |t|
-    t.bigint "language_preference_written", default: 0, null: false
-    t.bigint "language_preference_spoken", default: 0, null: false
+    t.integer "language_preference_written", default: 0, null: false
+    t.integer "language_preference_spoken", default: 0, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
